Dialogue and Fight Two
Adds all dialogue leading up to and following fight 2. Includes sequence
of finding new weapon, as well as a health potion - user has option to
completely ignore finding these.

TODO: need to figure out string formatting, I want all the dialogue to
land on the same part of the command prompt. In order to do that now,
all mt dialogue has to be on the same part of my code, which makes my
code logic very hard to follow.